requirements.txt
jupyter==1.0.0
matplotlib==3.2.1
networkx==2.4
numpy==1.18.1
pandas==1.0.2
pytket==0.5
pytket-pyzx==0.2.1
-e git+https://github.com/Quantomatic/pyzx.git@e214f1f681a8d352a09e5f9bef7f80ea3f3442de#egg=pyzx
